アプリを閉じたかどうかの判定を行った
