--- noPhone/noPhoneApp.swift
@@ -1,17 +1,0 @@
-//
-//  noPhoneApp.swift
-//  noPhone
-//
-//  Created by 藤本皇汰 on 2024/11/29.
-//
-
-import SwiftUI
-
-@main
-struct noPhoneApp: App {
-    var body: some Scene {
-        WindowGroup {
-            ContentView()
-        }
-    }
-}
